requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
python-dotenv>=0.19.0
google-cloud-firestore>=2.0.0
spotipy>=2.19.0
google-api-python-client>=2.0.0
google-auth-oauthlib>=0.4.0
urllib3<3.0.0
qdrant-client>=1.1.1
openai>=0.27.0
numpy>=1.21.0
pandas>=1.3.0
python-multipart>=0.0.5